refactor: Grouped related files (analyze-project)
