+++ src/steps/analyze-project/analyze-components/find-arguments.ts
@@ -1,7 +1,7 @@
 // import { AST as ASTJavaScript } from '@codemod-utils/ast-javascript';
 // import { AST as ASTTemplate } from '@codemod-utils/ast-template';
 
-import type { Signature } from '../../types/index.js';
+import type { Signature } from '../../../types/index.js';
 
 export function findArguments({
   classFile,
