+++ src/steps/analyze-project/analyze-components/index.ts
@@ -1,0 +1,3 @@
+export * from './find-arguments.js';
+export * from './find-blocks.js';
+export * from './find-element.js';
